Updated metadactyl access logging and exception handling
All exceptions are now thrown as a map containing [:type, :error, and
optional keys]. The optional keys are placed into the exception response
automatically (not nested in a ":fields" key).

Routes are the only functions which should return success responses.
All non-route functions have been cleansed of returning any kind of Ring
response.

Throwables and exceptions are no longer passed to the response logging
middlewares through the response map. The response logging middlewares
will only be executed if all is successful. It is now the reponsibility
of the exception handlers to log response errors.

CORE-7112
CORE-7113